# BE/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://project.test/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/todos": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Retrieve todos",
                "operationId": "18e676fc7a1838c48c63274e52066ac8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number for pagination",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "The title search todo",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "The start_date search for todo",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "due_date",
                        "in": "query",
                        "description": "The due_date search for todo",
                        "schema": {
                            "type": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Todo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "Store new Todo",
                "description": "Returns Todo data",
                "operationId": "storeTodo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TodoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Todo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/todos/{id}": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Get todo information",
                "description": "Returns todo data",
                "operationId": "gettodoById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "todo id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Todo"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Todos"
                ],
                "summary": "Update Todo",
                "description": "Returns Todo data",
                "operationId": "updateTodo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TodoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Todos"
                ],
                "summary": "Delete existing todo",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteTodo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "todo id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/assign-user": {
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "assign user to Todo",
                "description": "assign user to Todo",
                "operationId": "assignUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignTodoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/change-status": {
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "Change status to Todo",
                "description": "Change status to Todo",
                "operationId": "changeStatus",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeStatusTodoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AssignTodoRequest": {
                "title": "Assign user to Todo request",
                "description": "Assign user to Todo reques body data",
                "required": [
                    "user_id",
                    "todo_ids"
                ],
                "properties": {
                    "todo_ids": {
                        "title": "Due Date",
                        "description": "Due Date",
                        "type": "array",
                        "items": {
                            "type": "array",
                            "items": {}
                        },
                        "example": "[1,2]"
                    },
                    "user_id": {
                        "title": "Assignee ID",
                        "description": "Assignee's id of the new todo",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "ChangeStatusTodoRequest": {
                "title": "Change status to Todo request",
                "description": "Change status to Todo reques body data",
                "required": [
                    "todo_ids",
                    "status"
                ],
                "properties": {
                    "todo_ids": {
                        "title": "Due Date",
                        "description": "Due Date",
                        "type": "array",
                        "items": {
                            "type": "array",
                            "items": {}
                        },
                        "example": "[1,2]"
                    },
                    "status": {
                        "title": "Todo status",
                        "description": "Todo status",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "TodoRequest": {
                "title": "Store Todo request",
                "description": "Store Todo request body data",
                "required": [
                    "title",
                    "start_date",
                    "due_date"
                ],
                "properties": {
                    "title": {
                        "title": "Title",
                        "description": "title of the new todo",
                        "type": "string",
                        "example": "A nice todo"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the new todo",
                        "type": "string",
                        "example": "This is new todo's description"
                    },
                    "start_date": {
                        "title": "Start date",
                        "description": "Start date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27"
                    },
                    "due_date": {
                        "title": "Due Date",
                        "description": "Due Date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27"
                    },
                    "user_id": {
                        "title": "Assignee ID",
                        "description": "Assignee's id of the new todo",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "Todo": {
                "title": "Todo",
                "description": "Todo model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "Title",
                        "description": "title of the new todo",
                        "type": "string",
                        "example": "A nice todo"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the new todo",
                        "type": "string",
                        "example": "This is new todo's description"
                    },
                    "start_date": {
                        "title": "Start date",
                        "description": "Start date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27"
                    },
                    "due_date": {
                        "title": "Due Date",
                        "description": "Due Date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "Deleted at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "user_id": {
                        "title": "Assignee ID",
                        "description": "Assignee's id of the new todo",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Todo"
                }
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email of the new user",
                        "type": "string",
                        "example": "A nice todo"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of the new user",
                        "type": "string",
                        "example": "This is new todo's description"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "Deleted at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "todos": {
                        "title": "Todos",
                        "description": "Todo author's user model",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Todo"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        }
    },
...
}